# services/control-plane-api/openapi/control_plane_openapi_v1.4.yaml
openapi: 3.0.3
info:
  title: AI Digital Office OS Control Plane
  version: 1.4.0
  description: >
    Governed control-plane API for AI DIGITAL OFFICE OS. Every endpoint is
    tenant-scoped via the JWT 'tenant_id' claim, versioned under /v1, and
    returns the standard error envelope on any non-2xx response (see
    ErrorResponse schema). v1.4 closes the integrity-review findings against
    v1.3: full project/agent-version/run/deployment/usage/tool lifecycle
    coverage, an explicit auth contract (issuer/audience/JWKS/scopes), and
    a required Idempotency-Key header on every state-changing endpoint.
    This specification is product-agnostic: no endpoint, schema, or example
    is specific to any named downstream project (see blueprint clause 75).
servers:
  - url: https://api.digitaloffice.example/v1
    description: Production (placeholder host — real endpoint is an
      environment-specific choice, intentionally not fabricated here)
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        OAuth2 client-credentials token for service/agent callers, or
        OAuth2 authorization-code token for human users (blueprint clause 70).
        Required claims: 'iss' (issuer — this Office's or an approved
        external IdP), 'aud' (this control plane), 'exp'/'iat' (max lifetime
        15 minutes for service/agent callers, 8 hours for human sessions),
        'tenant_id', 'scope' (granted permission scopes), and 'actor_type'
        (SERVICE | HUMAN). Clock-skew tolerance ±60 seconds. JWKS rotates on
        a maximum 90-day cycle with a minimum 7-day overlap window. Session
        and refresh-token lifecycle is delegated entirely to the issuing IdP
        (blueprint clause 71) — this API only validates presented tokens.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string }
      description: >
        Required on every state-changing (POST/PATCH/DELETE) request.
        Replaying the same key within 24 hours returns the original result
        rather than reprocessing (blueprint clauses 49, 63, 68).
  schemas:
    ErrorResponse:
      type: object
      required: [error_code, message, correlation_id, retryable]
      properties:
        error_code:
          type: string
          enum: [VALIDATION_ERROR, AUTHORIZATION_ERROR, POLICY_ERROR, TOOL_ERROR,
                 MODEL_ERROR, PROVIDER_ERROR, NETWORK_ERROR, TIMEOUT,
                 DATABASE_ERROR, ARTIFACT_ERROR, DEPLOYMENT_ERROR, UNKNOWN_ERROR]
        message: { type: string }
        correlation_id: { type: string }
        retryable: { type: boolean }
    Task:
      type: object
      required: [task_id, tenant_id, status]
      properties:
        task_id: { type: string }
        tenant_id: { type: string, format: uuid }
        project_id: { type: string }
        workflow_id: { type: string }
        parent_task_id: { type: string }
        assigned_agent: { type: string }
        required_capability: { type: string }
        priority: { type: string }
        risk_level: { type: string, enum: [GREEN, YELLOW, RED] }
        security_level: { type: string }
        dependencies: { type: array, items: { type: string } }
        input: { type: object }
        expected_output: { type: object }
        status:
          type: string
          enum: [CREATED, QUEUED, ASSIGNED, RUNNING, WAITING, BLOCKED, FAILED, RETRYING, COMPLETED, CANCELLED, ESCALATED]
        retry_count: { type: integer }
        idempotency_key: { type: string }
        deadline: { type: string, format: date-time }
    TaskCreateRequest:
      type: object
      required: [project_id, required_capability, input, idempotency_key]
      properties:
        project_id: { type: string }
        workflow_id: { type: string }
        parent_task_id: { type: string }
        required_capability: { type: string }
        priority: { type: string, enum: [LOW, NORMAL, HIGH, CRITICAL] }
        risk_level: { type: string, enum: [GREEN, YELLOW, RED] }
        security_level: { type: string }
        dependencies: { type: array, items: { type: string } }
        input: { type: object }
        expected_output: { type: object }
        idempotency_key:
          type: string
          description: Required — guarantees a retried request never double-executes.
        deadline: { type: string, format: date-time }
...
    RoutingRequest:
      type: object
      required: [task_id, agent_id, required_capability]
      properties:
        task_id: { type: string }
        agent_id: { type: string }
        required_capability: { type: string }
        security_classification: { type: string }
        data_classification: { type: string }
        max_cost: { type: number }
        max_latency_ms: { type: integer }
    RoutingDecision:
      type: object
      properties:
        task_id: { type: string }
        agent_id: { type: string }
        selected_provider: { type: string }
        selected_model: { type: string }
        reason: { type: string }
        policy_result: { type: string, enum: [ALLOW, DENY, REQUIRE_APPROVAL, REQUIRE_ESCALATION] }
        fallback_chain_used: { type: array, items: { type: string } }
        decided_at: { type: string, format: date-time }
...
    Project:
      type: object
      required: [project_name, project_type]
      properties:
        project_id: { type: string }
        tenant_id: { type: string, format: uuid }
        project_name: { type: string }
        project_type:
          type: string
          description: "Free-form. No product type receives special-cased treatment (blueprint clause 75)."
        constitution_version: { type: string }
        lifecycle_state:
          type: string
          enum: [DISCOVERY, RESEARCH, REQUIREMENTS, PRODUCT_PLAN, ARCHITECTURE, UX_UI, DESIGN, DEVELOPMENT, CODE_REVIEW, QA, SECURITY, STAGING, APPROVAL, DEPLOYMENT, MONITORING, MAINTENANCE]
        risk_level: { type: string, enum: [GREEN, YELLOW, RED] }
        owner_user_id: { type: string, format: uuid }
        repository_ref: { type: string }
        environment_policy: { type: object }
...
    ModelEvaluationTriggerRequest:
      type: object
      required: [model_id, benchmark_suite]
      properties:
        model_id: { type: string }
        benchmark_suite: { type: string }
        evaluator_version: { type: string }
    ModelEvaluationResult:
      type: object
      properties:
        evaluation_id: { type: string, format: uuid }
        model_id: { type: string }
        benchmark_suite: { type: string }
        score: { type: number }
        metrics:
          type: array
          items:
            type: object
            properties:
              metric_name: { type: string }
              metric_value: { type: number }
              unit: { type: string }
        executed_at: { type: string, format: date-time }
...
  responses:
    Unauthorized:
      description: Missing/invalid bearer token
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
    PolicyDenied:
      description: Central Policy Engine returned DENY
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
    NotFound:
      description: Resource not found (or not visible to this tenant)
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } }
paths:
  /tasks:
    post:
      summary: Create a governed task
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TaskCreateRequest' }
      responses:
        '201':
          description: Task created
          content: { application/json: { schema: { $ref: '#/components/schemas/Task' } } }
        '403': { $ref: '#/components/responses/PolicyDenied' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /tasks/{task_id}:
    get:
      summary: Get task state
      parameters:
        - { name: task_id, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Task state
          content: { application/json: { schema: { $ref: '#/components/schemas/Task' } } }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /models/route:
    post:
      summary: Request capability-based model routing
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RoutingRequest' }
      responses:
        '200':
          description: Routing decision
          content: { application/json: { schema: { $ref: '#/components/schemas/RoutingDecision' } } }
        '403': { $ref: '#/components/responses/PolicyDenied' }
...
  /secrets/{reference_id}/rotate:
    post:
      summary: Trigger rotation of a referenced secret (never returns the raw value)
      parameters:
        - { name: reference_id, in: path, required: true, schema: { type: string, format: uuid } }
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '202': { description: Rotation triggered — audited as a secret-access event }
        '403': { $ref: '#/components/responses/PolicyDenied' }

  # ==================== New in v1.4 — closes the "incomplete API lifecycle" finding ====================
  /projects:
    post:
      summary: Create a project (authoritative project_registry record)
      parameters: [ { $ref: '#/components/parameters/IdempotencyKey' } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/Project' } } }
      responses:
        '201': { description: Project created, content: { application/json: { schema: { $ref: '#/components/schemas/Project' } } } }
    get:
      summary: List projects for the caller's tenant
      parameters:
        - { name: lifecycle_state, in: query, schema: { type: string } }
      responses:
        '200': { description: Project list, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Project' } } } } }
  /projects/{project_id}:
    get:
      summary: Get a project record
      parameters: [ { name: project_id, in: path, required: true, schema: { type: string } } ]
      responses:
        '200': { description: Project record, content: { application/json: { schema: { $ref: '#/components/schemas/Project' } } } }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      summary: Update a project's mutable fields (lifecycle_state, risk_level, environment_policy, etc.)
      parameters: [ { name: project_id, in: path, required: true, schema: { type: string } } ]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object } } }
      responses:
        '200': { description: Updated project, content: { application/json: { schema: { $ref: '#/components/schemas/Project' } } } }
...
  /models:
    get:
      summary: List registered models with capability tags
      parameters:
        - { name: capability, in: query, schema: { type: string } }
      responses:
        '200': { description: Model list, content: { application/json: { schema: { type: array, items: { type: object } } } } }
  /models/evaluate:
    post:
      summary: Trigger a model evaluation run
      parameters: [ { $ref: '#/components/parameters/IdempotencyKey' } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/ModelEvaluationTriggerRequest' } } }
      responses:
        '202': { description: Evaluation triggered, content: { application/json: { schema: { $ref: '#/components/schemas/ModelEvaluationResult' } } } }
  /models/evaluations:
    get:
      summary: List historical model evaluation runs (clause 64)
      parameters:
        - { name: model_id, in: query, schema: { type: string } }
      responses:
        '200': { description: Evaluation history, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/ModelEvaluationResult' } } } } }
...